:white_check_mark: First stab at header y center test
